refactor(obd2): decompose Obd2Service — fuel-rate chain, odometer chain, PSA listen-mode stream extracted (#3540)
1894 → 1317 lines. Three sibling collaborators following the #811/#3528
pattern: Obd2FuelRateReader (the mass/5E/MAF/speed-density chain + trim
correction + profile-η_v resolution, behind the narrow Obd2FuelRateReads
port the service implements), Obd2OdometerReader (primitives-only:
send tear-off + strategy string), and psaCanFrameStream (the #1418
ATCRA/STMA/STMP wiring). Service keeps thin delegators — public API and
behavior unchanged; file-length snapshot ratcheted down.

Closes #3540
